moved content of ExecOpRequestTPFatTPFServerWithTranslation to superclass
