+++ src/main/java/se/liu/ida/hefquin/engine/queryplan/executable/impl/ops/ExecOpRequestTPFatBRTPFServerWithTranslation.java
@@ -8,10 +8,12 @@ import se.liu.ida.hefquin.engine.federation.access.TriplePatternRequest;
  * This implementation handles pagination of the TPF; that is, it requests all
  * the pages, one after another.
  */
-public class ExecOpRequestTPFatBRTPFServerWithTranslation extends ExecOpRequestTPFWithTranslation<BRTPFServer>
+public class ExecOpRequestTPFatBRTPFServerWithTranslation extends ExecOpGenericTriplePatternRequestWithTranslation<BRTPFServer>
 {
 	public ExecOpRequestTPFatBRTPFServerWithTranslation( final TriplePatternRequest req, final BRTPFServer fm ) {
 		super( req, fm );
+		
+		assert fm.getVocabularyMapping() != null;
 	}
 	
 }
